transfer hook cli: refactor account meta parsing into new module
