+++ token/transfer-hook/cli/src/main.rs
@@ -1,4 +1,7 @@
+pub mod meta;
+
 use {
+    crate::meta::parse_transfer_hook_account,
     clap::{crate_description, crate_name, crate_version, Arg, Command},
     solana_clap_v3_utils::{
         input_parsers::{parse_url_or_moniker, pubkey_of_signer},
@@ -20,38 +23,8 @@ use {
         get_extra_account_metas_address,
         instruction::{initialize_extra_account_meta_list, update_extra_account_meta_list},
     },
-    std::{fmt, process::exit, rc::Rc, str::FromStr},
-    strum_macros::{EnumString, IntoStaticStr},
+    std::{process::exit, rc::Rc},
 };
-
-#[derive(Debug, Clone, Copy, PartialEq, EnumString, IntoStaticStr)]
-#[strum(serialize_all = "kebab-case")]
-pub enum AccountMetaRole {
-    Readonly,
-    Writable,
-    ReadonlySigner,
-    WritableSigner,
-}
-impl fmt::Display for AccountMetaRole {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self)
-    }
-}
-fn parse_transfer_hook_account(arg: &str) -> Result<AccountMeta, String> {
-    match arg.split(':').collect::<Vec<_>>().as_slice() {
-        [address, role] => {
-            let address = Pubkey::from_str(address).map_err(|e| format!("{e}"))?;
-            let meta = match AccountMetaRole::from_str(role).map_err(|e| format!("{e}"))? {
-                AccountMetaRole::Readonly => AccountMeta::new_readonly(address, false),
-                AccountMetaRole::Writable => AccountMeta::new(address, false),
-                AccountMetaRole::ReadonlySigner => AccountMeta::new_readonly(address, true),
-                AccountMetaRole::WritableSigner => AccountMeta::new(address, true),
-            };
-            Ok(meta)
-        }
-        _ => Err("Transfer hook account must be present as <ADDRESS>:<ROLE>".to_string()),
-    }
-}
 
 fn clap_is_valid_pubkey(arg: &str) -> Result<(), String> {
     is_valid_pubkey(arg)
